creating for Render deployment

+++ server/controllers/posts.js
@@ -29,8 +29,11 @@ export const createPost = async(req,res)=>{
 }
 
 export const deletePost = async(req,res)=>{
-    const id = req.params.id
+    const userId = req.userId
+    const id = req.params.id //postID
     if(!mongoose.Types.ObjectId.isValid(id)) return res.status(404).send('No post with that ID')
+    const currentPost = await PostMessage.findById(id)
+    if(currentPost.creator!==userId) return res.status(404).send('Unable to delete other\'s post')
     await PostMessage.findByIdAndRemove(id)
     //logging into console
     console.log(' request is made on url ',req.url,'with req type: ',req.method, 'reqest body is ',req.body,'id to delete is ',id)
@@ -46,13 +49,20 @@ export const updatePost=async(req,res)=>{
 
 }
 export const likePost =async(req,res) => {
-    const {id} = req.params
+    const {id} = req.params//postID
+    if (!req.userId)    return res.json({ message: "Unauthenticated" });
+    let userId  = req.userId
+
     console.log(' request is made on url ',req.url,'with req type: ',req.method, 'reqest body is ',req.body)
 
     if(!mongoose.Types.ObjectId.isValid(id)) return res.status(404).send('No post with that ID')
-
     let post = await PostMessage.findById(id)
-    const updatedPost = await PostMessage.findByIdAndUpdate(id, {likeCount:post.likeCount+1},{new:true})
+    const index = post.likes.indexOf(userId)
+    let updatedPost
+    if (index === -1) {
+        updatedPost = await PostMessage.findByIdAndUpdate(id,{likes:[...post.likes,userId]})
+        } else {updatedPost = await       PostMessage.findByIdAndUpdate(id,{likes:[...post.likes.filter(user=>user!==userId)]})}
+
     console.log(updatedPost)
     res.json(updatedPost)
 }
